refactor: Mudança das importações dos ícones
